Conflicts:
	.gitignore
	jeeves/src/main/java/jeeves/config/springutil/JeevesContextLoaderListener.java
	jeeves/src/main/java/jeeves/server/sources/http/JeevesServlet.java
	web-client/src/main/resources/apps/search/index.html
	web-client/src/main/resources/apps/search/index_debug.html
	web-client/src/main/resources/apps/tabsearch/index.html
	web-client/src/main/resources/apps/tabsearch/index_debug.html
	web/pom.xml
	web/src/main/webResources/WEB-INF/web.xml
	web/src/main/webapp/WEB-INF/config-security/config-security-mapping.xml
	web/src/main/webapp/xslt/base-layout-cssjs-loader.xsl

+++ harvesters/src/main/java/org/fao/geonet/kernel/harvest/harvester/AbstractHarvester.java
@@ -267,7 +267,7 @@ public abstract class AbstractHarvester<T extends HarvestResult> {
         for (String sourceUuid : sources) {
             Long ownedBySource = 
                     metadataRepository.count(Specifications.where(MetadataSpecs.hasSource(sourceUuid)));
-            if (ownedBySource == 0 && !sourceUuid.equals(params.uuid)) {
+            if (ownedBySource == 0 && !sourceUuid.equals(params.uuid) && sourceRepository.exists(sourceUuid)) {
                 removeIcon(sourceUuid);
                 sourceRepository.delete(sourceUuid);
             }
@@ -643,7 +643,7 @@ public abstract class AbstractHarvester<T extends HarvestResult> {
     /**
      * @return
      */
-    public AbstractParams getParams() {
+    public synchronized AbstractParams getParams() {
         return params;
     }
 
@@ -658,9 +658,9 @@ public abstract class AbstractHarvester<T extends HarvestResult> {
      * @throws SQLException
      */
     protected void doDestroy() throws SQLException {
-        removeIcon(params.uuid);
+        removeIcon(getParams().uuid);
 
-        context.getBean(SourceRepository.class).delete(params.uuid);
+        context.getBean(SourceRepository.class).delete(getParams().uuid);
         // FIXME: Should also delete the categories we have created for servers
     }
 
@@ -846,7 +846,7 @@ public abstract class AbstractHarvester<T extends HarvestResult> {
         el.addContent(new Element(name).setText(Integer.toString(value)));
     }
 
-    public void setParams(AbstractParams params) {
+    public synchronized void setParams(AbstractParams params) {
         this.params = params;
     }
